requirements.txt
Office365-REST-Python-Client>=2.3.1
python-dotenv>=0.19.1
